:recycle: ImgCarousel 컴포넌트화
